+++ src/debugger/manageddebugger.h
@@ -120,6 +120,7 @@ class ManagedDebugger
                           std::vector<GotoTarget> &targets, std::string &output);
     HRESULT Goto(ThreadId threadId, uint32_t targetId, std::string &output);
     HRESULT GetEmbeddedSource(const Source &source, std::string &sourceContent);
+    void GetLoadedSources(std::vector<Source> &sources);
 
     void WriteStdin(gsl::span<const char> text);
     bool InitializeRemoteConsoleServer(int port);