+++ src/protocol/dapio.cpp
@@ -225,6 +225,7 @@ void DAPIO::AddCapabilitiesTo(json &capabilities)
     capabilities.emplace("supportsLogPoints", true);
     capabilities.emplace("supportsGotoTargetsRequest", true);
     capabilities.emplace("supportsSingleThreadExecutionRequests", true);
+    capabilities.emplace("supportsLoadedSourcesRequest", true);
 }
 
 void DAPIO::SetupProtocolLogging(const std::string &path)